examples/broadcast-bot/requirements.txt
olvid-bot>=1.0.0
aiohttp
